Split test into different files
